refactor: Give custom exceptions own module
