+++ anonymizer.py
@@ -1,5 +1,5 @@
 import util
-from util import FormException
+from exceptions import FormException
 
 import xml.etree.ElementTree as ET
